final/requirements.txt
asgiref @ file:///C:/ci_311/asgiref_1676448296195/work
certifi @ file:///C:/ci_311/certifi_1676421629969/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django @ file:///C:/ci_311/django_1676469589007/work
django-cors-headers==3.14.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse @ file:///C:/ci_311/sqlparse_1676456541280/work
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
uritemplate==4.1.1
urllib3==1.26.15
wincertstore==0.2
